Start working on culling vertices outside of convex hull
Since some of the non-rectangular shapes can end up creating points outside of the expected area, the resulting vertices need to be pruned by the bounds of the object. My plan is to see if the vertex is found to be behind and of the planes of the solid if each of the planes are facing towards the center of the object. Theoretically it should work, but I think I goofed the maths.