Cull vertices outside of convex hull, fix irregular shapes, clean up, and more!
I'm so happy this is finally working.

This commit should be the final step in converting brush geometry. Now all regular, irregular, and brush entities should be properly converted *with correct normals*!
Also a bunch of cleanup of old code and comments.

Plugging in a VMF will now get you a greybox brush blockout of the map!