Fix displacement texture support
Now displacements will have their correct textures! Though, it is important to note that if a displacement has a blend texture, only the default texture will be shown -- i.e there will be no blending. The data is included in the displacement object, but in order to implement it into an obj object, either a separate texture must be generated with the blend built in, or a per-vertex material is applied and blended between.
The downsides of the first option is that it requires a lot of processing before hand (and Java's compatibility with Targa files is kinda potato at best), and the resulting texture is not editable.
The downside of the second option is that the each vertex has to be either entirely one texture or entirely the other, and the linear blend will always go between the two. This will make for very strange looking textures when comparing it to the hammer counterpart, and it would require a complete rework of how the application handles textures.

The perfect-world solution is to use a format that supports this kind of thing out of the box, but then you loose compatibility pretty quickly.

I think I'm going to keep it like this for now, just using the base texture.