Add prop_* support!
Huzzah! This took so much time, but I'm pretty happy with where it stands right now. As an overview, Props will be extracted from the supplied VPK, decompiled using crowbar, it's QC deserialized and analysed before finally converting the SMD to OBJ format, taking acount for Source's fun coordinate and rotation system. Each of the textures used by the model are also found and converted just like brush materials.
This is a huge step because most other programs that deal with the VMF format straight up ignore models and either pretend they don't exist or put a placeholder object and hope the user can do the work themselves. This application will actually decompile the model and it's materials and place in the right location with the right rotation. If it isn't obvious, I'm really happy with this improvement.

It's not perfect yet though, there's still a lot that needs to be done. Namely, a lot of garbage is left behind from the decompiling stages (but there still is from the material converting as well, I'll get to it eventually), and this only amplifies the need for custom resource support.

The next commit will clean up a lot of WIP things so I can get organized again.