Add support for vertex normals and fixed textures with 8bit alphas
Vertex normals from decompiled props will now be loaded in correctly, and decompiling textures was improved.

Many new VTF textures contain 8bit alpha channels, even if the texture itself does not have any discernible transparency. TGA and PNG files seem to choke and die (or maybe its the VTFLib conversion, not sure) when given these layers (again kinda, it's possible to fish them out, and I was able to see them correctly in blender when fiddling with alpha settings), but because it caused more adverse results that things that looked good, the application now looks to see if the VMT material uses transparency. If it does, the VTF is converted to a TGA, as before. Otherwise, the VTF is converted to a JPG, without any alpha channel whatsoever.