Admit defeat for infodecal and info_overlay
I spent a real long time trying to think of a solution to the problem, but in the end I can't think of something that would work better than brute forcing the problem. Maybe I can write that solution out once the project can be called stable, but there are still a couple things that need to get done before then.