Add support for external resources, fix a bunch of problems
Finally things are starting to work again. This commit adds support for resources like models or textures to be added from the filesystem, rather than just from the VPK. This is necessary for any map with custom content, such as many user-made maps.
This also includes a ton of small fixes, like not hanging when numbers in scientific notation are found within the VMF, dashes within models names, or whitespace padding. This also fixes a ton of edge cases, but I need to go to sleep now.