Move functions into classes
The application was build as if I were programming in a procedural language rather than an object oriented one. This commit moves functions out of the main class and into classes that make more sense, i.e. the function that parses VMF files has been moved to the VMF class.
There's still a lot of functions, and the main class is still gigantic, but for now I think it's ok, especially since the conversion process is long and sequential, there's a lot of calculations to do.