v1.0.0
I think the project is finally ready to be called complete! There are still things that can be improved, but all of the features I wanted to get done are done. This commit also bumps the version of the app and updates the readme to get rid of old development information