Fix Displacements!!
It hurts that I got so close for so long and find the solution just days after making the project public, but I am so so happy this works now. This bug has plagued me for months.

The `startposition` attribute isn't necessarily on a line with one of the sides of the brush, but it is placed nearest the point that is deemed the "bottom-left" corner of the displacement. Orienting all the other point around this makes everything line up beautifully.

Resolves Issue #6