Fix Crowbar subfolder setting causing problems
Crowbar has a setting called `DecompileFolderForEachModelIsChecked` that when set to true will cause decompiled models to always be placed in a subfolder with the model's name. While this setting is false by default, if it is true no models would be imported successfully.

Huge thanks to @Devostated for helping track down this issue