Move all temporary files to system temp folder
All files that are not needed after the completion of the program are now placed in a folder the the system's temp directory, and automatically deleted with a shutdown hook. This makes the application look cleaner since the working directory isn't polluted with files the user doesn't need to worry about.