Start working on UI
About time I got around to this. This commit begins the work on adding an actual user interface for the application. The interface is greatly inspired by bspsrc, as the process for decompiling from bsp to vmf is pretty similar to the user as converting from vmf to obj. The CLI is still used when running from the terminal, but the new gui is used whenever double clicking on the jar. I believe it's functional right now, but there's still a lot of testing, and currently the log only appears in the terminal, regardless of if it actually exists or not. That's the next thing to work on.

Also this commit changes the way parameters are computed. I'm not sure this is the final format, but it sure is nicer than the original.

(I hate Java UI programming)