Fix sorting colinear and equidistant points
Finally! This issue was brought up awhile ago but I was never able to figure it out. Instead of calculating the determinate to figure out the order (which fails for colinear points), and calculating the distance from the center (which fails on equidistant points, this find two perpendicular vectors in the plane given by the normal vector and calculates the triple product of the vector in question. This way it's order is not based off of just the determinate and will always be unique no matter if two vectors are colinear or equidistant from each other.

Resolves #14