Collapse almost duplicate vertices
If a vertex is within 0.2 units of an already calculated vertex on the same side, don't add it to the list. Similarly, if a vertex is within 0.2 units of an already calculated vertex of a different side of the same brush, copy it's exact location so they'll merge in a later step.
This greatly reduces the number of extraneous vertices when dealing with complex brushes. This does not effect entities at all.
Resolves #3