Fix displacements for real this time
Some reworks of the logic surrounding how displacement vertices are calculated, including taking account of irregular shapes and simplifying the process a little bit. When calculating the texture coordinates, I tried to put the calculation in a loop, however because the four points need to be in a specific order, I decided to just keep it expanded for now and try to tackle it later, since right now it works it's just a little ugly in the code.

Resolves #16
Resolves #20