Remove achievement arrays
Some games have logic_achievement entities that can award part of an achievement, such as the "Door Prize" achievement in Portal two, where each of six doors award part of the achievement through an array, such as `ACH.A3_DOORS[6]` for door number 6. This array breaks the parser, crashing the conversion. This commit removes the array portion of the achievement name, but keeps the rest in tact, just in case.

Resolves #25